requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
python-multipart==0.0.9
email-validator==2.1.1
password-strength==0.0.3.post2
bcrypt==4.1.2
playwright==1.41.2
httpx==0.27.0
pytest==8.2.1
pytest-asyncio==0.23.5
